move AnyDecoder/Encoder into separate file
